Created first selenium test (tests Title of page and logging in as a test user). Need to add selenium remote server set up into CI/CD pipeline, might need to wget, java selenium.jar, etc. manually in CircleCI workflow.
